// src/properties/enums/property.enums.ts
export enum PropertyOperationType {
  VENTA = 'VENTA',
  ALQUILER = 'ALQUILER',
  ALQUILER_TEMPORAL = 'ALQUILER_TEMPORAL',
}

export enum PropertyType {
  CASA = 'CASA',
  DEPARTAMENTO = 'DEPARTAMENTO',
  LOCAL = 'LOCAL',
  TERRENO = 'TERRENO',
  OFICINA = 'OFICINA',
  COCHERA = 'COCHERA',
  PH = 'PH',
  QUINTA = 'QUINTA',
}

export enum Currency {
  ARS = 'ARS',
  USD = 'USD',
  EUR = 'EUR',
}

export enum PropertyStatus {
  BORRADOR = 'BORRADOR',
  PUBLICADA = 'PUBLICADA',
  RESERVADA = 'RESERVADA',
  VENDIDA = 'VENDIDA',
  ALQUILADA = 'ALQUILADA',
  INACTIVA = 'INACTIVA',
}

export enum AvailableService {
  AGUA = 'agua',
  LUZ = 'luz',
  GAS = 'gas',
  CLOACAS = 'cloacas',
  CORDON_CUNETA = 'cordon_cuneta',
  PAVIMENTO = 'pavimento',
}